Remove duplicate rule requireSpecificAttributes from rules.md
The auto-generated `rules.md` had a duplicate rule in its description
for `requireSpecificAttributes`. The bug was due to the fact that an
unrelated rule `requireSpacesInsideAttributeBrackets` had an erroneous
duplicate description for the unrelated rule.

After removing the duplicate description and running `npm run docs`,
or more accurately `npm test`, the auto-generated `rules.md` file no
longer contains the duplicate rule.